[Rooms] Add room deletion cleanup job
When rooms are created a task is created to request their deletion
at a point in the future. Currently the timeout is 30 seconds.

In the future when there are mechanisms to enable players to join the room
we will want to cancel the task and then reschedule it in the event that
all players have left the room and it is empty